employee management system
Employee Management Project is to allow the administrator of any organization to edit and find out the personal details of an employee and allows the employee to keep up to date with his profile. It’ll also facilitate keeping all the records of employees, such as their id, name, mailing address, phone number, DOB, etc. So all the information about an employee will be available in a few seconds.